+++ preload.js
@@ -6,6 +6,7 @@ const { contextBridge } = electron;
 const { ipcRenderer } = electron;
 
 const deserializeJson = (serializedJavascript) => eval('(' + serializedJavascript + ')'); //deserialize function for 'serialize-javascript' library
+let onPreferencesUpdatedHandler; 
 
 contextBridge.exposeInMainWorld('api', {
 	getSections: () => deserializeJson(ipcRenderer.sendSync('getSections')),
@@ -15,5 +16,9 @@ contextBridge.exposeInMainWorld('api', {
 	setPreferences: preferences => ipcRenderer.send('setPreferences', preferences),
 	showOpenDialog: dialogOptions => ipcRenderer.sendSync('showOpenDialog', dialogOptions),
 	sendButtonClick: channel => ipcRenderer.send('sendButtonClick', channel),
-  encrypt: secret => ipcRenderer.sendSync('encrypt', secret),
+	encrypt: secret => ipcRenderer.sendSync('encrypt', secret),
+	onPreferencesUpdated: handler => { onPreferencesUpdatedHandler = handler; },
+});
+ipcRenderer.on('preferencesUpdated', (e, preferences) => {
+	typeof onPreferencesUpdatedHandler === "function" && onPreferencesUpdatedHandler(preferences);
 });